Update examples 2 and 3

+++ examples/example2.cuh
@@ -14,7 +14,7 @@ void example2_kernel(int* stack, int stacksize, long long* opstack, int opstacks
     F* valuestack, int valuestacksize, double* outputstack, int outputstacksize, int Nthreads) 
 {
     int s_size    = stacksize;
-    long long op_size  = opstacksize;
+    int op_size  = opstacksize;
     int v_size    = valuestacksize;
     int ou_size   = outputstacksize;
