+++ examples/example3.cuh
@@ -24,7 +24,7 @@ void example3_kernel(int* stack, int stacksize, long long* opstack, int opstacks
     F* valuestack, int valuestacksize, double* outputstack, int outputstacksize, int Nthreads) 
 {
     int s_size    = stacksize;
-    long long op_size  = opstacksize;
+    int op_size  = opstacksize;
     int v_size    = valuestacksize;
     int ou_size   = outputstacksize;
